+++ sdk/cosmos/azure-cosmos/test/test_resource_id.py
@@ -19,20 +19,16 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-
-import unittest
-
 import uuid
 import pytest
 import test_config
-import azure.cosmos.cosmos_client as cosmos_client
-from azure.cosmos.partition_key import PartitionKey
+from azure.cosmos import CosmosClient, PartitionKey
 
 pytestmark = pytest.mark.cosmosEmulator
 
 
 @pytest.mark.usefixtures("teardown")
-class ResourceIdTests(unittest.TestCase):
+class TestResourceIds:
     configs = test_config._test_config
     host = configs.host
     masterKey = configs.masterKey
@@ -40,17 +36,18 @@ class ResourceIdTests(unittest.TestCase):
     last_headers = []
 
     @classmethod
-    def setUpClass(cls):
+    def _set_up(cls):
         if (cls.masterKey == '[YOUR_KEY_HERE]' or
                 cls.host == '[YOUR_ENDPOINT_HERE]'):
             raise Exception(
                 "You must specify your Azure Cosmos account values for "
                 "'masterKey' and 'host' at the top of this class to run the "
                 "tests.")
-        cls.client = cosmos_client.CosmosClient(cls.host, cls.masterKey, connection_policy=cls.connectionPolicy)
-        cls.databaseForTest = cls.configs.create_database_if_not_exist(cls.client)
+        cls.client = CosmosClient(cls.host, cls.masterKey)
+        cls.created_database = cls.client.create_database_if_not_exists(test_config._test_config.TEST_DATABASE_ID)
 
     def test_id_unicode_validation(self):
+        self._set_up()
         # unicode chars in Hindi for Id which translates to: "Hindi is the national language of India"
         resource_id1 = u'हिन्दी भारत की राष्ट्रीय भाषा है'  # cspell:disable-line
 
@@ -58,31 +55,32 @@ class ResourceIdTests(unittest.TestCase):
         resource_id2 = "!@$%^&*()-~`'_[]{}|;:,.<>"
 
         # verify that databases are created with specified IDs
-        created_db1 = self.client.create_database(resource_id1)
-        created_db2 = self.client.create_database(resource_id2)
+        created_db1 = self.client.create_database_if_not_exists(resource_id1)
+        created_db2 = self.client.create_database_if_not_exists(resource_id2)
 
-        self.assertEqual(resource_id1, created_db1.id)
-        self.assertEqual(resource_id2, created_db2.id)
+        assert resource_id1 == created_db1.id
+        assert resource_id2 == created_db2.id
 
         # verify that collections are created with specified IDs
-        created_collection1 = created_db1.create_container(
+        created_collection1 = created_db1.create_container_if_not_exists(
             id=resource_id1,
             partition_key=PartitionKey(path='/id', kind='Hash'))
-        created_collection2 = created_db2.create_container(
+        created_collection2 = created_db2.create_container_if_not_exists(
             id=resource_id2,
             partition_key=PartitionKey(path='/id', kind='Hash'))
 
-        self.assertEqual(resource_id1, created_collection1.id)
-        self.assertEqual(resource_id2, created_collection2.id)
+        assert resource_id1 == created_collection1.id
+        assert resource_id2 == created_collection2.id
 
-        # verify that collections are created with specified IDs
-        item1 = created_collection1.create_item({"id": resource_id1})
-        item2 = created_collection1.create_item({"id": resource_id2})
+        # verify that items are created with specified IDs
+        item1 = created_collection1.upsert_item({"id": resource_id1})
+        item2 = created_collection1.upsert_item({"id": resource_id2})
 
-        self.assertEqual(resource_id1, item1.get("id"))
-        self.assertEqual(resource_id2, item2.get("id"))
+        assert resource_id1 == item1.get("id")
+        assert resource_id2 == item2.get("id")
 
-    def test_create_illegal_characters(self):
+    def test_create_illegal_characters_async(self):
+        self._set_up()
         database_id = str(uuid.uuid4())
         container_id = str(uuid.uuid4())
         partition_key = PartitionKey(path="/id")
@@ -91,8 +89,7 @@ class ResourceIdTests(unittest.TestCase):
         created_container = created_database.create_container(id=container_id, partition_key=partition_key)
 
         # Define errors returned by checks
-        illegal_chars_string = 'Id contains illegal chars.'
-        space_chars_string = 'Id ends with a space.'
+        error_strings = ['Id contains illegal chars.', 'Id ends with a space.']
 
         # Define illegal strings
         illegal_strings = [
@@ -107,30 +104,27 @@ class ResourceIdTests(unittest.TestCase):
         ]
 
         # test illegal resource id's for all resources
-        error_string = illegal_chars_string
         for resource_id in illegal_strings:
-            if resource_id == "ID_with_trailing_spaces   ":
-                error_string = space_chars_string
             try:
                 self.client.create_database(resource_id)
-                self.fail("Database create should have failed for id {}".format(resource_id))
+                pytest.fail("Database create should have failed for id {}".format(resource_id))
             except ValueError as e:
-                self.assertEquals(str(e), error_string)
+                assert str(e) in error_strings
 
             try:
                 created_database.create_container(id=resource_id, partition_key=partition_key)
-                self.fail("Container create should have failed for id {}".format(resource_id))
+                pytest.fail("Container create should have failed for id {}".format(resource_id))
             except ValueError as e:
-                self.assertEquals(str(e), error_string)
+                assert str(e) in error_strings
 
             try:
                 created_container.create_item({"id": resource_id})
-                self.fail("Item create should have failed for id {}".format(resource_id))
+                pytest.fail("Item create should have failed for id {}".format(resource_id))
             except ValueError as e:
-                self.assertEquals(str(e), error_string)
+                assert str(e) in error_strings
             try:
                 created_container.upsert_item({"id": resource_id})
-                self.fail("Item upsert should have failed for id {}".format(resource_id))
+                pytest.fail("Item upsert should have failed for id {}".format(resource_id))
             except ValueError as e:
-                self.assertEquals(str(e), error_string)
+                assert str(e) in error_strings